+++ lib/index.d.ts
@@ -7,3 +7,4 @@ export * from "./response";
 export * from "./router";
 export * from "./utils";
 export * from "./logger";
+export * from "./socket";