+++ src/api.ts
@@ -1,9 +1,17 @@
-import { ApiError, HttpError, MethodError, RouteError } from "./errors";
-import { HttpStatus } from "./http";
+import {
+  ApiError,
+  HttpError,
+  MethodError,
+  RouteError,
+  ActionError,
+} from "./errors";
+import { HttpIntegration, HttpStatus } from "./http";
 import { Logger, LoggerOptions, LogHandlerFunction } from "./logger";
+import { HandlerFunction } from "./options";
 import { ApiRequest } from "./request";
 import { ApiResponse, Response, ResponseError } from "./response";
-import { getRouteParams, HandlerFunction, Route, RouteParams } from "./router";
+import { getRouteParams, instanceOfRoute, Route, RouteParams } from "./router";
+import { Action, instanceOfAction, isDefaultAction } from "./socket";
 
 export interface ApiOptions {
   logger?: ApiOptionsLogger;
@@ -16,6 +24,7 @@ export interface ApiOptionsLogger {
 }
 export class Api {
   private routes: Route[] = [];
+  private actions: Action[] = [];
   private middlewares: HandlerFunction[] = [];
   private middlewareFinally?: HandlerFunction;
 
@@ -23,12 +32,15 @@ export class Api {
 
   constructor(private options: ApiOptions = {}) {}
 
-  use(middlewares: Route[] | HandlerFunction[]): Api {
+  use(middlewares: Route[] | Action[] | HandlerFunction[]): Api {
     for (const m of middlewares) {
+      /* istanbul ignore else */
       if (typeof m === "function") {
         this.middlewares.push(m);
-      } else {
+      } else if (instanceOfRoute(m)) {
         this.routes.push(m);
+      } else if (instanceOfAction(m)) {
+        this.actions.push(m);
       }
     }
 
@@ -43,6 +55,7 @@ export class Api {
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   async listen(event: any, context: any): Promise<Response> {
     const request = new ApiRequest(event, context);
+    const response = new ApiResponse(request);
     const loggerOptions: LoggerOptions = { context };
     const isTrace = this.options.logger?.trace || false;
 
@@ -56,39 +69,28 @@ export class Api {
     if (isTrace) this.trace("REQUEST", "request", request.toRequest());
 
     try {
-      if (!this.routes.length) {
-        throw new RouteError(request.path);
+      let responseHandler: any;
+
+      if (request.proxyIntegration === HttpIntegration.APIGW_WS_API) {
+        const action = this.findAction(request);
+
+        await this.executeMiddlewares(request);
+
+        responseHandler = await action.handler(request, response);
+      } else {
+        const [route, params] = this.findRoute(request);
+
+        await this.executeMiddlewares(request);
+
+        request.params = params;
+        responseHandler = await route.handler(request, response);
       }
 
-      const matchingMethodRoutes = this.routes.filter(
-        (r) => r.method === request.method
-      );
-
-      if (matchingMethodRoutes.length === 0) {
-        throw new MethodError(request.method, request.path);
-      }
-
-      const routeMatched = this.matchRoute(matchingMethodRoutes, request.path);
-
-      if (!routeMatched) {
-        throw new RouteError(request.path);
-      }
-
-      const response = new ApiResponse(request);
-      const [route, params] = routeMatched;
-
-      request.params = params;
-
-      await this.executeMiddlewares(request, response);
-
-      const responseHandler = await route.handler(request, response);
       const responseApi = response.toResponse();
+      const result = responseApi || response.send(responseHandler);
 
       /* istanbul ignore else */
-      if (this.middlewareFinally)
-        await this.middlewareFinally(request, response);
-
-      const result = responseApi || response.send(responseHandler);
+      if (this.middlewareFinally) await this.middlewareFinally(request, response);
 
       if (isTrace) this.trace("RESPONSE", "response", result);
 
@@ -108,10 +110,50 @@ export class Api {
     this.log.trace(message);
   }
 
-  private async executeMiddlewares(
-    request: ApiRequest,
-    response: ApiResponse
-  ): Promise<any> {
+  private findRoute(request: ApiRequest): [Route, RouteParams] {
+    if (!this.routes.length) {
+      throw new RouteError(request.path);
+    }
+
+    const matchingMethodRoutes = this.routes.filter(
+      (r) => r.method === request.method
+    );
+
+    if (matchingMethodRoutes.length === 0) {
+      throw new MethodError(request.method, request.path);
+    }
+
+    const routeMatched = this.matchRoute(matchingMethodRoutes, request.path);
+
+    if (!routeMatched) {
+      throw new RouteError(request.path);
+    }
+
+    return routeMatched;
+  }
+
+  private findAction(request: ApiRequest): Action {
+    if (!this.actions.length) {
+      throw new ActionError(request.route);
+    }
+
+    const requestAction = request.body.action || "";
+    const actionMatched = this.actions.find((a) => {
+      if (isDefaultAction(request.route)) {
+        return a.name === requestAction;
+      }
+      return a.name === request.route;
+    });
+
+    if (!actionMatched) {
+      throw new ActionError(request.route);
+    }
+
+    return actionMatched;
+  }
+
+  private async executeMiddlewares(request: ApiRequest): Promise<any> {
+    const response = new ApiResponse(request);
     for (let i = 0; i < this.middlewares.length; i++) {
       const middleware = this.middlewares[i];
 
@@ -164,6 +206,13 @@ export class Api {
         responseError.status = status;
         responseError.data = {
           path: err.path,
+        };
+      } else if (err instanceof ActionError) {
+        status = HttpStatus.NOT_FOUND;
+
+        responseError.status = status;
+        responseError.data = {
+          name: err.name,
         };
       }
     }