+++ src/http.ts
@@ -59,6 +59,7 @@ export enum HttpIntegration {
   ALB = "elb",
   APIGW_REST_API = "apigw-rest-api",
   APIGW_HTTP_API = "apigw-http-api",
+  APIGW_WS_API = "apigw-ws-api",
 }
 
 export const HttpHeadersCors = {