+++ src/options.ts
@@ -1,3 +1,6 @@
+import { ApiRequest } from "./request";
+import { ApiResponse } from "./response";
+
 export interface RouterOptions {
   basePath?: string;
   version?: string;
@@ -11,3 +14,8 @@ export interface CorsOptions {
   methods?: string;
   origin?: string;
 }
+
+export declare type HandlerFunction = (
+  request: ApiRequest,
+  response: ApiResponse
+) => void | any | Promise<any>;