+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [1.2.0] - 2020-02-20
+
+### Added
+- The component `socket` is added.
+
+### Changed
+- Fixed middleware response by problem with router response.
+
 ## [1.1.3] - 2020-02-16
 
 ### Changed
@@ -60,6 +68,7 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 ### Added
 - Structure project
 
+[1.2.0]: https://github.com/ldiego73/micro-lambda-api/releases/tag/1.2.0
 [1.1.3]: https://github.com/ldiego73/micro-lambda-api/releases/tag/1.1.3
 [1.1.2]: https://github.com/ldiego73/micro-lambda-api/releases/tag/1.1.2
 [1.1.1]: https://github.com/ldiego73/micro-lambda-api/releases/tag/1.1.1